requirements.txt
torch>=1.0.0
torchvision>=0.4.0
opencv-python
visdom
dominate
numpy
scipy
pillow<=9.5.0
unidecode
requests
tqdm
dlib
mediapipe
pytk
